Add memdump options to k8s configs.
When back-in-time aka memdump is enabled, it requires a larger /dev/shm
filesystem than is normally provided in k8s. To work around this, in the
pod specification create a memory-backed volume and mount it to
/dev/shm.

This allows for having an agent configuration that is eventually
compatible with back-in-time captures so it doesn't have to be changed later.