Daemonset yaml for split agent containers
Add variants of our v1/v2 daemonset yaml docs that support separate
containers for the steps of building the kernel module and running the
agent.

The container that builds the kernel module is specified as an init
container, which runs first and must be successful for the other
containers to run. The init container is sysdig/agent-kmodule and has
necessary mounts to build the kernel module.

The main container changes from sysdig/agent to sysdig/agent-slim, and
removes the mounts that were required to build the kernel module.

I tested both on a single-node k8s cluster running 1.6 and they ran as
expected.