Better k8s audit config (#88)
* Use default agent k8s audit port of 7765.

* Improved docs/scripts for many K8s distros

Revamp our instructions/supporting scripts to document how to enable K8s
Audit Support for a wider set of K8s Distributions:

* minishift-3.11
* openshift-3.11
* openshift-4.2
* gke
* eks
* iks
* rke-1.13: Implies K8s >= 1.13
* kops
* minikube-1.13: Implies K8s >= 1.13
* minikube-1.12: Implies K8s 1.11/1.12

For each distribution, include step-by-step instructions to show how to
enable K8s audit support. Also revamp the enable_k8s_audit.sh script to
handle each of the above distributions, other than eks.

The README.md is suitable for conversion to our official docs, but it
can be used on an unofficial basis for now.

I'm also including a audit-log-forwarder.yaml file, using fluent-bit,
that can be used to generically read a log file and forward the contents
to a webhook. I didn't end up using it, but it may come in handy for
other distributions where you can emit audit events to a log file only.

* Fetch stackdriver-webhook-bridge.yaml from github

The image hasn't been pushed yet.

* Save service account keys to pwd not home

Other files are also created there so it's a good precedent.

* Create audit sink yaml for openshift 4.2

* Fix typo (not minikube vm)

* Remove detailed docs from README.md

The official docs will be the single place for the instructions. This
directory only holds the yaml files and scripts.