Remove hard sudo requirement
The script will fail immediately with the message "Please install sudo and re-run the script" if sudo is not present in the shell, like in some Docker containers. Sudo is only required for installing curl if it is not present, and since the [official IBM docs recommend downloading the script via curl](https://cloud.ibm.com/docs/services/Monitoring-with-Sysdig?topic=Sysdig-config_agent#config_agent_linux), most users won't ever hit that codepath.