[get_support_bundle.sh] Add safeguard when getting containers name from pod
It can happen that a pod listed in the previous steps is not present anymore when it comes its time during the for loop (CLBOs, eviction, etc.).
This causes the script to crash since it tries to access the non existen pod to retrieve the list of containers.

This commit is adding a safeguard for this kind of situation.