Add build ID to support bundle
This PR aims to add a simple build number that we can use to identify which version of the support bundle is being used. I'm proposing that we use a build ID that contains the date of the PR in the format of yyyymmdd. We will need to remember to update this when changes are made. If there is a better way to do this, please let me know.